requirements.txt
# requirements.txt
flask>=2.0.1
accelerate
gunicorn>=20.1.0
diffusers
torch
pillow
numpy
langchain-openai
python-dotenv
transformers
sentence_transformers
pinecone
regex
wikipedia
wikipediaapi
langgraph
langchain-community
langchain-core
